ユーザー検索機能 Draft

+++ src/main/resources/db/migration/V1.0.0__CREATE_USERS.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    id INT PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(255),
     email VARCHAR(255),
     phone VARCHAR(255),
